Shored up expect function test logic.
The expect_exit_code argument in expect
is now checked properly against the exit code
of the function being tested, rather than arbitrary output
from the function itself.

Prior to this change:

expect echo "1" 1

would return true as the output of echo "1" is... 1

but the exit code of echo "1" is 0, which is what we really
want to be tested.

So now

expect echo "1" 1

will fail as the echo command successfully executed and didn't send
a non-zero exit code.